+++ src/render/apiError.ts
@@ -1,0 +1,6 @@
+import moveList from '../components/MoiveList';
+
+export default () => {
+  moveList.removeSkeleton();
+  moveList.showErrorMessage();
+};
