+++ src/render/skeleton.ts
@@ -1,0 +1,5 @@
+import moveList from '../components/MoiveList';
+
+export default {
+  create: () => moveList.createSkeleton,
+};
